slashcommands-handler-with-mqtt/requirements.txt
paho-mqtt~=1.2.1
